requirements.txt
streamlit>=1.32.0
google-generativeai>=0.3.2
Pillow>=10.0.0
python-dotenv>=1.0.0
typing-extensions>=4.5.0